refactor: 금액 관련 자료형 모두 long으로 변경

+++ src/main/java/lotto/domain/LottoMachine.java
@@ -11,7 +11,7 @@ public class LottoMachine {
 
     public LottoMachine(PurchaseAmount purchaseAmount,
                         LottoNumberGenerator lottoNumberGenerator) {
-        this.lottoCount = purchaseAmount.getAmount() / LOTTO_PRICE;
+        this.lottoCount = (int)(purchaseAmount.getAmount() / LOTTO_PRICE);
         this.lottoNumberGenerator = lottoNumberGenerator;
     }
